+++ CHANGELOG.md
@@ -6,6 +6,45 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+
+### Changed
+
+### Removed
+
+## [0.0.2-dev] - 2021.06.20
+
+### Added
+
+* hostname, organization and token can now be set with the respective environment values to align with [TFE Provider](https://registry.terraform.io/providers/hashicorp/tfe/latest/docs). ([#7](https://github.com/straubt1/tfx/issues/7))
+  * TFE_HOSTNAME
+  * TFE_ORGANIZATION
+  * TFE_TOKEN
+* Added "message" flag to `tfx run` and `tfx plan` commands. ([#8](https://github.com/straubt1/tfx/issues/8))
+* `tfx workspace` commands
+  * `list` - List all workspaces in an Organization (optional workspace name search string) 
+  * `list all` - List all workspaces in All Organizations the API token has access to (optional workspace name search string) 
+  * `show` - Show details of a workspace
+* `tfx metrics` command to pull details about TFx (this command is hidden)
+  * Organization Count
+  * Workspace Count
+  * Run Count
+  * Policy Check Count
+  * Policies Pass/Fail Count
+* `tfx metrics workspace` command to get run metrics for all workspaces in a single organization (this command is hidden)
+  * Can filter on start date
+  * Output:
+    * Workspace Name
+    * Total Runs
+    * Errored Runs
+    * Discarded Runs
+    * Cancelled Runs
+
+### Changed
+
+* Cleaned up docs
+
+### Removed
 
 ## [0.0.1-dev] - 2021.05.23